docs/requirements.txt
requests
pyyaml
Jinja2
